refactor: move update submission API call to store
